models: widen entity ids i32 -> i64 to match sea-orm 2.0 BIGINT PKs
The loco 0.16->1.1 / sea-orm 1.1->2.0 upgrade changed auto-increment
primary keys from INTEGER to BIGINT: loco's ColType::PkAuto now maps to
big_pk_auto, and its create_table foreign-key helper emits BigInteger to
match the referenced 64-bit PK. The generated entity structs, however,
still declared id/*_id as i32, so every query decoding an id against a
freshly-migrated Postgres failed with:

  Query Error: decoding column "id": Rust type Option<i32> (INT4) is not
  compatible with SQL type INT8

which surfaced as HTTP 500 on /api/auth/current the first time a tenant was
promoted onto a real Postgres, sending the SPA into a sign-in redirect loop.

Why it hid until now:
- the test suite runs on SQLite, which is loosely typed (INT4/INT8 are
  interchangeable), so the mismatch never tripped in CI;
- the deployment had no live tenant DBs (canaries route to a static backend,
  no castle app, no castle schema), so the sea-orm-2.0 migrations had never
  run against Postgres until the first promote.

Fix: widen every PK and FK to i64 across entities, model helpers, controller
Path/params, views, and test fixtures - exactly what loco 1.1's entity
generator produces today. No migration change needed (the columns are already
BIGINT); this only aligns the Rust side. Stale "INTEGER NOT NULL, FK" comments
in the migrations corrected to BIGINT.

Verified against a throwaway Postgres 18.6 (production's version): the full
request-test suite - which does INSERT ... RETURNING id and reads rows back
into the entities, the exact operation that 500'd - passes 11/11, where it
fails on the old i32 entities.

+++ src/controllers/comments.rs
@@ -19,7 +19,7 @@ pub struct CreateCommentParams {
 async fn load_viewable_finding(
     ctx: &AppContext,
     user: &users::Model,
-    finding_id: i32,
+    finding_id: i64,
 ) -> Result<findings::Model> {
     let finding = findings::Entity::find_by_id(finding_id)
         .one(&ctx.db)
@@ -46,7 +46,7 @@ async fn load_viewable_finding(
 #[debug_handler]
 pub async fn list(
     CurrentUser(user): CurrentUser,
-    Path(finding_id): Path<i32>,
+    Path(finding_id): Path<i64>,
     State(ctx): State<AppContext>,
 ) -> Result<Response> {
     let finding = load_viewable_finding(&ctx, &user, finding_id).await?;
@@ -68,7 +68,7 @@ pub async fn list(
 #[debug_handler]
 pub async fn create(
     CurrentUser(user): CurrentUser,
-    Path(finding_id): Path<i32>,
+    Path(finding_id): Path<i64>,
     State(ctx): State<AppContext>,
     Json(params): Json<CreateCommentParams>,
 ) -> Result<Response> {
