+++ src/models/_entities/project_members.rs
@@ -9,10 +9,10 @@ pub struct Model {
     pub created_at: DateTimeWithTimeZone,
     pub updated_at: DateTimeWithTimeZone,
     #[sea_orm(primary_key)]
-    pub id: i32,
+    pub id: i64,
     pub role: String,
-    pub project_id: i32,
-    pub user_id: i32,
+    pub project_id: i64,
+    pub user_id: i64,
 }
 
 #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
